Bug fixes and Quality of Life improvements
Fixes:
* fixes a bug where an evt_ handler's requested propegation delay was
ignored if the time requested was a float instead of an int
* Prevents "known modes" from being added to the known mode notification
lists more than once (e.g,. if you create them multiple times in your
code)
Improvements:
* PDLEDs can now be scheduled using the schedule command -- thanks Josh!
* Attempting to play a lampshow that is not registered will now generate
a warning instead of crashing the game
* The game can continue running if an asset is missing!
For animation, define an asset with the key 'missing' in the asset list
under the Animations section
For fonts define an asset with the key 'default'
Note: of course, things may not look right, but you will get errors in
the log and can continue testing and add the missing assets later.
* if no ballsave lamp exists, game can continue
* dropped trough ball settle time to 0.25s
* tags may be used in some places instead of specific switch/lamp names
in the machine yaml.  Examples:
- Trough switches may either be named trough? (where ? is a number or
letter) or they may be tagged: tag: trough
- The shoot_again lamp may have any name if given the tag: shoot_again
* the quick_load parameter to the asset_manager has been changed in
favor of an option to pass in a specific asset file or asset dictionary.
This is useful for quick launching a game without having to load your
entire asset file, but also letting you customize which assets are being
loaded depending on what modes you are trying to test.  An example of
this is forthcoming.
Event changes:
* evt_player_added will be called on _any_ modes, not just active ones.
(to support this, skeletonGame.notifyModes() now supports a parameter
`only_active_modes` which is True by default)
* Added a new event: `evt_shoot_again`, which will be fired after
evt_ball_ending, but before evt_ball_starting if the player has an extra
ball.  This gives your modes the opportunity to show a SHOOT AGAIN
message, or play a lampshow, etc. -- because the mode that handles this
might no longer be in the queue because of the ball over at this time,
this is also broadcast to all modes (active and inactive)