Trough, Outhole and early_save improvements
1. For outhole trough style support: In your machine yaml, tag a switch
(or name it) 'outhole' and do the same for a coil.  The trough will
automatically add code to pulse the outhole when occupied for 300ms.

2. If you don't want to name your shooter lane switch 'shooter' just add
a tag in the machine yaml with the value 'shooter'

3. early_save switches are now supported.  Add an `early_save` tag for a
switch in the machine yaml (`tag: early_save`) and that switch will
automatically queue up a ball if it is closed when the ball saver is
active.  Note: don't use this on your left outlane if you have a
kickback!  Only use it on switches that will always lead to a drain
(e.g., right outlane).