switch to weakrefs in known_modes
Experiences PyProcGame programmers are used to re-creating their modes
in reset() to ensure a clean state.  While we can debate the utility of
doing so,  manually re-creating advancedModes in the game's reset()
method was causing the previous instances to also stick around.
Switching to weakrefs fixed that so that you can instantiate
AdvancedModes in the game's init() or reset()