Error catching for reset() bugs
This change will raise an Exception when a user (developer) tries to
destroy and recreate an AdvancedMode in reset().  Doing this causes
problems for the framework.  You're recommended to create the
AdvancedModes once (in __init__ or another mode's mode_init() or other
methods that are called only once).   The issue is that SkeletonGame
knows when a mode is created, but setting the variable to None is not
the same thing as deleting it.  As a result, the old mode is still known
to the framework (even despite the use of weakref).  Easier to not do
this, and there's a performance benefit to not doing so.  I recommend
you define a .reset() method if you want to clear your variables based
on a machine "reset" occurance instead of evt_player_added or
evt_game_ending.