rigorous reconciliation of game_user_settings file
loading a user settings file makes a very concerted effort to ensure the
settings file is "sane".
* The old version added missing values using the defaults described in
the settings template: `game_default_settings.yaml`
* This version also makes sure that there are no extra sections or items
in the user's `game_user_settings.yaml` and, for values selected from a
list, it ensures the user's settings file contains a valid value from
that list (the default value will be provided if not).
**values in numeric ranges may be outside the numeric range.