fontsyle on layer creation
you can now pass in a fonstyle when creating an HDfont layer vs having
to set the style in a second command.  You can still pass in a style
when wen doing a set_tex to temporarily over rid the set style, you can
also still permentaly change the default style for the layer.

If you pass in style, it will ignore any other style attribute that is
passed in (fill color, interior color, line_color and line width)